res/res_config_odbc: Fix improper escaping of backslashes with MySQL
When escaping backslashes with MySQL, the proper way to escape the characters
in a LIKE clause is to escape the '\' four times, i.e., '\\\\'. To quote the
MySQL manual:

"Because MySQL uses C escape syntax in strings (for example, ?\226?\128?\156\n?\226?\128?\157 to represent
a newline character), you must double any ?\226?\128?\156\?\226?\128?\157 that you use in LIKE strings.
For example, to search for ?\226?\128?\156\n?\226?\128?\157, specify it as ?\226?\128?\156\\n?\226?\128?\157. To search for ?\226?\128?\156\?\226?\128?\157,
specify it as ?\226?\128?\156\\\\?\226?\128?\157; this is because the backslashes are stripped once by the
parser and again when the pattern match is made, leaving a single backslash to
be matched against."

ASTERISK-24808 #close
Reported by: Javier Acosta
patches:
  res_config_odbc.diff uploaded by Javier Acosta (License 6690)
........

Merged revisions 432720 from http://svn.asterisk.org/svn/asterisk/branches/11
........

Merged revisions 432721 from http://svn.asterisk.org/svn/asterisk/branches/13
........

Merged revisions 432722 from file:///srv/subversion/repos/asterisk/trunk