Replace the internal PJSIP resolver implementation at a lower level.
Unfortunately there's no easy way to cover all cases where resolution occurs
without a change to PJSIP. I've made such a change which allows the resolution
process to be taken over by an external implementation. This change has been
sent to Teluu for inclusion and I will work with them to ensure it is.

As a result though... no changes are required elsewhere to perform resolution
using the new DNS work. It's just happening.

This also fixes a bug with resolution of an IP address where the addresses
structure was filled with garbage.