Streamline the allocation and sorting functions for SRV and NAPTR.
This places allocation and sorting into a table, thus allowing simpler
lookups of allocation and sorting by record type. Additional record types
can be added by adding an appropriate row in the allocation and sorting
tables.

This may get taken a step further by having record types register themselves
with the DNS core, but for the time being this seems like overkill since
all DNS types exist in the core, and not in loadable modules.